Fixes issues #18, #19, #20, #21, and #22
This update to the script does the following:
Allows the script to use rad.tif files if atmocorr.tif files are not available in order to create the refl.tif files. Also renames the atmcorr_files and atmcorr_count to refl_ready_images and refl_ready_count, respectively.
Removes some remnant lines from rad.py that are not relevant to the creation of refl.tif files
Renames the rad variable in now line 170 to refl.
Prints a message indicating which image was used to successfully create a refl.tif file.