Fixes issue #36
(Hopefully) fixes the issue where the script wouldn't detect the output .txt files of atmcorr_regr.py.
This script also relied on the name of the folder to look for the name of the output .txt.  This was changed so that it now looks for each image's associated xml file to find the the name of the output .txt.