Starts addressing issue #44
Start to the classification parameter script based on previous structures. Mostly ripped from the refl script, with mostly name adjustments and extraneous lines removed. The new function executed on the image has been adjusted to carry out a sum all bands parameter. This parameter will lead to three classes based on relatively arbitrary thresholds. The output should be a raster of values associated with each of the three classes.