Reverted in-script writing location, adjusted thresholding, added check for nanometer vs micrometer
The script now does all of the image writing at the end to ensure that it'll be more easily changed in the future.
The abundances are now normalized in the script (the largest sum of the endmember abundances out of each pixel is divided out of every abundance value).
The script now turns raw imagery in nanometers into micrometers to be more in line with the images converted into reflectance.